adapt unit tests
