+++ process/eventsHandler_test.go
@@ -6,6 +6,7 @@ import (
 	"testing"
 
 	"github.com/multiversx/mx-chain-core-go/core/check"
+	"github.com/multiversx/mx-chain-core-go/data/block"
 	"github.com/multiversx/mx-chain-core-go/data/outport"
 	"github.com/multiversx/mx-chain-core-go/data/smartContractResult"
 	"github.com/multiversx/mx-chain-core-go/data/transaction"
@@ -13,15 +14,23 @@ import (
 	"github.com/multiversx/mx-chain-notifier-go/data"
 	"github.com/multiversx/mx-chain-notifier-go/mocks"
 	"github.com/multiversx/mx-chain-notifier-go/process"
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
 func createMockEventsHandlerArgs() process.ArgsEventsHandler {
 	return process.ArgsEventsHandler{
-		CheckDuplicates:      false,
-		Locker:               &mocks.LockerStub{},
+		Locker: &mocks.LockerStub{
+			HasConnectionCalled: func(ctx context.Context) bool {
+				return true
+			},
+			IsEventProcessedCalled: func(ctx context.Context, blockHash string) (bool, error) {
+				return true, nil
+			},
+		},
 		Publisher:            &mocks.PublisherStub{},
 		StatusMetricsHandler: &mocks.StatusMetricsStub{},
+		EventsInterceptor:    &mocks.EventsInterceptorStub{},
 	}
 }
 
@@ -61,6 +70,17 @@ func TestNewEventsHandler(t *testing.T) {
 		require.Nil(t, eventsHandler)
 	})
 
+	t.Run("nil events interceptor", func(t *testing.T) {
+		t.Parallel()
+
+		args := createMockEventsHandlerArgs()
+		args.EventsInterceptor = nil
+
+		eventsHandler, err := process.NewEventsHandler(args)
+		require.Equal(t, process.ErrNilEventsInterceptor, err)
+		require.Nil(t, eventsHandler)
+	})
+
 	t.Run("should work", func(t *testing.T) {
 		t.Parallel()
 
@@ -71,8 +91,274 @@ func TestNewEventsHandler(t *testing.T) {
 	})
 }
 
+func TestHandleSaveBlockEvents(t *testing.T) {
+	t.Parallel()
+
+	t.Run("duplicated events, should return early", func(t *testing.T) {
+		t.Parallel()
+
+		args := createMockEventsHandlerArgs()
+		args.CheckDuplicates = true
+
+		args.Locker = &mocks.LockerStub{
+			IsEventProcessedCalled: func(ctx context.Context, blockHash string) (bool, error) {
+				return false, nil
+			},
+		}
+
+		args.EventsInterceptor = &mocks.EventsInterceptorStub{
+			ProcessBlockEventsCalled: func(eventsData *data.ArgsSaveBlockData) (*data.InterceptorBlockData, error) {
+				require.Fail(t, "should have not been called")
+				return nil, nil
+			},
+		}
+
+		eventsHandler, err := process.NewEventsHandler(args)
+		require.Nil(t, err)
+
+		err = eventsHandler.HandleSaveBlockEvents(data.ArgsSaveBlockData{})
+		require.Nil(t, err)
+	})
+
+	t.Run("failed to pre-process events, should fail", func(t *testing.T) {
+		t.Parallel()
+
+		args := createMockEventsHandlerArgs()
+		args.CheckDuplicates = true
+
+		args.Locker = &mocks.LockerStub{
+			IsEventProcessedCalled: func(ctx context.Context, blockHash string) (bool, error) {
+				return true, nil
+			},
+		}
+
+		expectedErr := errors.New("expected err")
+		args.EventsInterceptor = &mocks.EventsInterceptorStub{
+			ProcessBlockEventsCalled: func(eventsData *data.ArgsSaveBlockData) (*data.InterceptorBlockData, error) {
+				return nil, expectedErr
+			},
+		}
+
+		eventsHandler, err := process.NewEventsHandler(args)
+		require.Nil(t, err)
+
+		err = eventsHandler.HandleSaveBlockEvents(data.ArgsSaveBlockData{})
+		require.Equal(t, expectedErr, err)
+	})
+
+	t.Run("should work", func(t *testing.T) {
+		t.Parallel()
+
+		blockHash := "blockHash1"
+		txs := map[string]*outport.TxInfo{
+			"hash1": {
+				Transaction: &transaction.Transaction{
+					Nonce: 1,
+				},
+				ExecutionOrder: 1,
+			},
+		}
+		scrs := map[string]*outport.SCRInfo{
+			"hash2": {
+				SmartContractResult: &smartContractResult.SmartContractResult{
+					Nonce: 2,
+				},
+			},
+		}
+		logData := []*outport.LogData{
+			{
+				Log: &transaction.Log{
+					Address: []byte("logaddr1"),
+					Events:  []*transaction.Event{},
+				},
+				TxHash: "logHash1",
+			},
+		}
+
+		logEvents := []data.Event{
+			{
+				Address: "addr1",
+			},
+		}
+
+		header := &block.HeaderV2{
+			Header: &block.Header{
+				ShardID: 2,
+			},
+		}
+		blockData := data.ArgsSaveBlockData{
+			HeaderHash: []byte(blockHash),
+			TransactionsPool: &outport.TransactionPool{
+				Transactions:         txs,
+				SmartContractResults: scrs,
+				Logs:                 logData,
+			},
+			Header: &block.HeaderV2{},
+		}
+
+		expTxs := map[string]*transaction.Transaction{
+			"hash1": {
+				Nonce: 1,
+			},
+		}
+		expScrs := map[string]*smartContractResult.SmartContractResult{
+			"hash2": {
+				Nonce: 2,
+			},
+		}
+
+		expTxsData := data.BlockTxs{
+			Hash: blockHash,
+			Txs:  expTxs,
+		}
+		expScrsData := data.BlockScrs{
+			Hash: blockHash,
+			Scrs: expScrs,
+		}
+		expLogEvents := data.BlockEvents{
+			Hash:    blockHash,
+			Events:  logEvents,
+			ShardID: 2,
+		}
+
+		expTxsWithOrder := map[string]*outport.TxInfo{
+			"hash1": {
+				Transaction: &transaction.Transaction{
+					Nonce: 1,
+				},
+				ExecutionOrder: 1,
+			},
+		}
+		expScrsWithOrder := map[string]*outport.SCRInfo{
+			"hash2": {
+				SmartContractResult: &smartContractResult.SmartContractResult{
+					Nonce: 2,
+				},
+			},
+		}
+		expTxsWithOrderData := data.BlockEventsWithOrder{
+			Hash:    blockHash,
+			ShardID: 2,
+			Txs:     expTxsWithOrder,
+			Scrs:    expScrsWithOrder,
+			Events:  logEvents,
+		}
+
+		pushWasCalled := false
+		txsWasCalled := false
+		scrsWasCalled := false
+		blockEventsWithOrderWasCalled := false
+
+		args := createMockEventsHandlerArgs()
+
+		args.EventsInterceptor = &mocks.EventsInterceptorStub{
+			ProcessBlockEventsCalled: func(eventsData *data.ArgsSaveBlockData) (*data.InterceptorBlockData, error) {
+				return &data.InterceptorBlockData{
+					Hash:          blockHash,
+					Header:        header,
+					Txs:           expTxs,
+					Scrs:          expScrs,
+					LogEvents:     logEvents,
+					TxsWithOrder:  expTxsWithOrder,
+					ScrsWithOrder: expScrsWithOrder,
+				}, nil
+			},
+		}
+
+		args.Publisher = &mocks.PublisherStub{
+			BroadcastCalled: func(events data.BlockEvents) {
+				pushWasCalled = true
+				assert.Equal(t, expLogEvents, events)
+			},
+			BroadcastTxsCalled: func(event data.BlockTxs) {
+				txsWasCalled = true
+				assert.Equal(t, expTxsData, event)
+			},
+			BroadcastScrsCalled: func(event data.BlockScrs) {
+				scrsWasCalled = true
+				assert.Equal(t, expScrsData, event)
+			},
+			BroadcastBlockEventsWithOrderCalled: func(event data.BlockEventsWithOrder) {
+				blockEventsWithOrderWasCalled = true
+				assert.Equal(t, expTxsWithOrderData, event)
+			},
+		}
+
+		eventsHandler, err := process.NewEventsHandler(args)
+		require.Nil(t, err)
+
+		err = eventsHandler.HandleSaveBlockEvents(blockData)
+		require.Nil(t, err)
+
+		assert.True(t, pushWasCalled)
+		assert.True(t, txsWasCalled)
+		assert.True(t, scrsWasCalled)
+		assert.True(t, blockEventsWithOrderWasCalled)
+	})
+}
+
+func TestShouldProcessSaveBlockEvents(t *testing.T) {
+	t.Parallel()
+
+	t.Run("should process", func(t *testing.T) {
+		t.Parallel()
+
+		args := createMockEventsHandlerArgs()
+		args.CheckDuplicates = true
+
+		args.Locker = &mocks.LockerStub{
+			IsEventProcessedCalled: func(ctx context.Context, blockHash string) (bool, error) {
+				return true, nil
+			},
+		}
+
+		eventsHandler, err := process.NewEventsHandler(args)
+		require.Nil(t, err)
+
+		shouldProcess := eventsHandler.ShouldProcessSaveBlockEvents("blockHash1")
+		require.True(t, shouldProcess)
+	})
+
+	t.Run("duplicated events, should not process", func(t *testing.T) {
+		t.Parallel()
+
+		args := createMockEventsHandlerArgs()
+		args.CheckDuplicates = true
+
+		args.Locker = &mocks.LockerStub{
+			IsEventProcessedCalled: func(ctx context.Context, blockHash string) (bool, error) {
+				return false, nil
+			},
+		}
+
+		eventsHandler, err := process.NewEventsHandler(args)
+		require.Nil(t, err)
+
+		shouldProcess := eventsHandler.ShouldProcessSaveBlockEvents("blockHash1")
+		require.False(t, shouldProcess)
+	})
+}
+
 func TestHandlePushEvents(t *testing.T) {
 	t.Parallel()
+
+	t.Run("empty hash should return error", func(t *testing.T) {
+		t.Parallel()
+
+		args := createMockEventsHandlerArgs()
+		eventsHandler, err := process.NewEventsHandler(args)
+		require.Nil(t, err)
+
+		events := data.BlockEvents{
+			Hash:      "",
+			ShardID:   1,
+			TimeStamp: 1234,
+			Events:    []data.Event{},
+		}
+
+		err = eventsHandler.HandlePushEvents(events)
+		require.Equal(t, common.ErrReceivedEmptyEvents, err)
+	})
 
 	t.Run("broadcast event was called", func(t *testing.T) {
 		t.Parallel()
@@ -96,43 +382,9 @@ func TestHandlePushEvents(t *testing.T) {
 		eventsHandler, err := process.NewEventsHandler(args)
 		require.Nil(t, err)
 
-		eventsHandler.HandlePushEvents(events)
-		require.True(t, wasCalled)
-	})
-
-	t.Run("check duplicates enabled, should not process event", func(t *testing.T) {
-		t.Parallel()
-
-		blockEvents := data.BlockEvents{
-			Hash:      "hash1",
-			ShardID:   1,
-			TimeStamp: 1234,
-			Events:    []data.Event{},
-		}
-
-		wasCalled := false
-		args := createMockEventsHandlerArgs()
-		args.CheckDuplicates = true
-		args.Publisher = &mocks.PublisherStub{
-			BroadcastCalled: func(events data.BlockEvents) {
-				require.Equal(t, blockEvents, events)
-				wasCalled = true
-			},
-		}
-		args.Locker = &mocks.LockerStub{
-			IsEventProcessedCalled: func(ctx context.Context, blockHash string) (bool, error) {
-				return false, nil
-			},
-		}
-
-		eventsHandler, err := process.NewEventsHandler(args)
+		err = eventsHandler.HandlePushEvents(events)
 		require.Nil(t, err)
-
-		shouldProcess := eventsHandler.ShouldProcessPushEvents(blockEvents.Hash)
-		require.False(t, shouldProcess)
-
-		eventsHandler.HandlePushEvents(blockEvents)
-		require.False(t, wasCalled)
+		require.True(t, wasCalled)
 	})
 }
 
@@ -278,40 +530,6 @@ func TestHandleTxsEvents(t *testing.T) {
 		eventsHandler.HandleBlockTxs(blockTxs)
 		require.True(t, wasCalled)
 	})
-
-	t.Run("check duplicates enabled, should not process event", func(t *testing.T) {
-		t.Parallel()
-
-		blockTxs := data.BlockTxs{
-			Hash: "hash1",
-			Txs: map[string]*transaction.Transaction{
-				"hash1": {
-					Nonce: 1,
-				},
-			},
-		}
-
-		wasCalled := false
-		args := createMockEventsHandlerArgs()
-		args.CheckDuplicates = true
-		args.Publisher = &mocks.PublisherStub{
-			BroadcastTxsCalled: func(event data.BlockTxs) {
-				require.Equal(t, blockTxs, event)
-				wasCalled = true
-			},
-		}
-		args.Locker = &mocks.LockerStub{
-			IsEventProcessedCalled: func(ctx context.Context, blockHash string) (bool, error) {
-				return false, nil
-			},
-		}
-
-		eventsHandler, err := process.NewEventsHandler(args)
-		require.Nil(t, err)
-
-		eventsHandler.HandleBlockTxs(blockTxs)
-		require.False(t, wasCalled)
-	})
 }
 
 func TestHandleScrsEvents(t *testing.T) {
@@ -343,39 +561,6 @@ func TestHandleScrsEvents(t *testing.T) {
 
 		eventsHandler.HandleBlockScrs(blockScrs)
 		require.True(t, wasCalled)
-	})
-
-	t.Run("check duplicates enabled, should not process event", func(t *testing.T) {
-		t.Parallel()
-
-		wasCalled := false
-		args := createMockEventsHandlerArgs()
-		args.CheckDuplicates = true
-		args.Publisher = &mocks.PublisherStub{
-			BroadcastScrsCalled: func(event data.BlockScrs) {
-				wasCalled = true
-			},
-		}
-		args.Locker = &mocks.LockerStub{
-			IsEventProcessedCalled: func(ctx context.Context, blockHash string) (bool, error) {
-				return false, nil
-			},
-		}
-
-		eventsHandler, err := process.NewEventsHandler(args)
-		require.Nil(t, err)
-
-		events := data.BlockScrs{
-			Hash: "hash1",
-			Scrs: map[string]*smartContractResult.SmartContractResult{
-				"hash2": {
-					Nonce: 2,
-				},
-			},
-		}
-
-		eventsHandler.HandleBlockScrs(events)
-		require.False(t, wasCalled)
 	})
 }
 
@@ -411,31 +596,6 @@ func TestHandleBlockEventsWithOrderEvents(t *testing.T) {
 
 		eventsHandler.HandleBlockEventsWithOrder(events)
 		require.True(t, wasCalled)
-	})
-
-	t.Run("check duplicates enabled, should not process event", func(t *testing.T) {
-		t.Parallel()
-
-		wasCalled := false
-		args := createMockEventsHandlerArgs()
-		args.CheckDuplicates = true
-		args.Publisher = &mocks.PublisherStub{
-			BroadcastBlockEventsWithOrderCalled: func(event data.BlockEventsWithOrder) {
-				require.Equal(t, events, events)
-				wasCalled = true
-			},
-		}
-		args.Locker = &mocks.LockerStub{
-			IsEventProcessedCalled: func(ctx context.Context, blockHash string) (bool, error) {
-				return false, nil
-			},
-		}
-
-		eventsHandler, err := process.NewEventsHandler(args)
-		require.Nil(t, err)
-
-		eventsHandler.HandleBlockEventsWithOrder(events)
-		require.False(t, wasCalled)
 	})
 }
